refactor(cli): split RTDB worker client by responsibility
