+++ packages/cli/src/serve/worker/client.ts
@@ -49,7 +49,39 @@ export * from './client/firestore-reads.js';
 export * from './client/firestore-writes.js';
 export * from './client/rules.js';
 export * from './client/admin-firestore.js';
-export * from './client/rtdb.js';
+export { rtdbGetDatabase, rtdbRef, rtdbChild } from './client/rtdb-references.js';
+export {
+  rtdbGet,
+  rtdbSet,
+  rtdbSetPriority,
+  rtdbSetWithPriority,
+  rtdbUpdate,
+  rtdbRemove,
+  rtdbPush,
+} from './client/rtdb-operations.js';
+export {
+  rtdbOnValue,
+  rtdbOnChildAdded,
+  rtdbOnChildChanged,
+  rtdbOnChildRemoved,
+  rtdbOnChildMoved,
+  rtdbOff,
+} from './client/rtdb-listeners.js';
+export {
+  rtdbRunTransaction,
+  type RtdbTransactionOptions,
+  type RtdbTransactionResult,
+} from './client/rtdb-transactions.js';
+export {
+  RtdbOnDisconnect,
+  rtdbOnDisconnect,
+  rtdbGoOffline,
+  rtdbGoOnline,
+} from './client/rtdb-connection-lifecycle.js';
+export {
+  rtdbServerTimestamp,
+  rtdbConnectDatabaseEmulator,
+} from './client/rtdb-controls.js';
 export * from './client/rtdb-admin.js';
 export * from './client/studio.js';
 export * from './client/presence.js';
