+++ packages/cli/src/serve/worker/client/rtdb-controls.ts
@@ -1,0 +1,8 @@
+/** RTDB environment controls and server values. */
+export function rtdbServerTimestamp(): { readonly __rtdbSentinel: 'serverTimestamp' } {
+  return { __rtdbSentinel: 'serverTimestamp' };
+}
+
+export function rtdbConnectDatabaseEmulator(): void {
+  // Shared worker sandbox is already local.
+}
